update server start functions
this changes the way udpbd-vexfat and udpbd-server are run.
The default option with show console unchecked will start the server as a background process.
The reason for this change is because of "QuckEdit mode".
If a user interacts with the server console, the server hangs which causes many issues including game freezes or failure to launch games.