bugfix for udpbd cold boot
This fixes a bug where on the first launch of a game via udpbd, udpbd-vexfat throws this error:
Failed to reply with UDPBD_CMD_INFO_REPLY to 192.168.91.40:48573: A socket operation was attempted to an unreachable network. (os error 10051)
This error was reproduceable by restarting the pc that is running the server.
The error was thrown only once per restart of the pc.
It seems that loading the dev9 irx file and waiting a couple seconds then running the neutrino elf fixes the issue.